Move `setup` to `db setup` as well
